move vars.js and fix its imports
