+++ data/dbConfig.js
@@ -2,14 +2,14 @@
 // use this file for other db wrappers or orm connectors
 const knex = require('knex')
 const knexfile = require('../knexfile')
-const { DB_ENV } = require('../vars')
-const knexDbConnection = knex(knexfile[DB_ENV])
+const vars = require('../utils/vars')
+const knexDbConnection = knex(knexfile[vars.get('DB_ENV')])
 
 const db = {
-  knex: knexDbConnection
+  knex: knexDbConnection,
 }
 
-function getConnection(connection = "knex") {
+function getConnection(connection = 'knex') {
   return db[connection]
 }
